+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fixed blocksCreate, blocksDestroy, usersLookup, usersShow, usersSearch to reflect version 1.1 of the API.
 * Implemented usersContributees, usersContributors.
+* Implemented friendshipsLookup, friendshipsUpdate, friendsList, followersList.
 
 # Changelog since 2.2.1